+++ application/backend/src/main/resources/libs/bulma/components/common/rte/configurations/default/.content.json
@@ -17,6 +17,7 @@
     "icon": "code",
     "plugin": {
       "sling:resourceType": "extensions/dialogs/components/richtext/plugin/code"
-    }
+    },
+    "sling:orderBefore": "bulletlist"
   }
 }
