--- application/extensions/bulma-code/README.md
@@ -1,1 +1,0 @@
-# RTE Code extension
