--- application/extensions/bulma-code/package.json
@@ -1,33 +1,0 @@
-{
-  "scripts": {
-    "lint": "eslint --ignore-path .eslintignore .",
-    "lintfix": "eslint --ignore-path .eslintignore --fix .",
-    "babel": "babel ./src/main/webapp --extensions \".ts,.tsx\" --config-file ./babel/.babelrc.js -d target/dist --copy-files",
-    "watch": "babel ./src/main/webapp/ --config-file ./babel/.babelrc.js --extensions \".ts,.tsx\" -d target/dist --copy-files --watch",
-    "prepare": "snowpack --dest target/dist/apps/bulma-code/web-resources --optimize"
-  },
-  "dependencies": {
-    "@tiptap/extension-code": "^2.0.3",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "snowpack": "^1.5.1"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.20.7",
-    "@babel/core": "7.7.7",
-    "@babel/preset-env": "^7.18.2",
-    "@babel/preset-react": "7.8.3",
-    "@babel/preset-typescript": "^7.16.0",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "babel-plugin-module-resolver": "^5.0.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
-    "concurrently": "5.1.0",
-    "cross-env": "^7.0.2",
-    "csstype": "^3.1.1",
-    "rollup-plugin-node-polyfills": "0.2.1",
-    "rollup-plugin-postcss": "2.6.4",
-    "serve": "^14.1.2",
-    "type-fest": "^3.4.0",
-    "typescript": "^4.5.2"
-  }
-}
