--- application/extensions/bulma-color/README.md
@@ -1,1 +1,0 @@
-# RTE Color extension
