--- application/extensions/bulma-color/babel/import-map.json
@@ -1,8 +1,0 @@
-{
-  "rootDir": "/apps/",
-  "aliases": {
-    "^(components.*)": "websight-dialogs-view/web-resources/\\1.js",
-    "^(.*)": "websight-atlaskit-esm/web-resources/\\1.js"
-  },
-  "devAliases": {}
-}
