--- application/extensions/bulma-color/bnd.bnd
@@ -1,1 +1,0 @@
-Sling-Bundle-Resources: /libs/extensions/dialogs/components/richtext;/libs/bulma-color
