--- application/extensions/bulma-color/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "include": [
-    "../src/main/webapp/**/*"
-  ],
-  "compilerOptions": {
-    "baseUrl": "./bulma-color",
-    "lib": [
-      "esnext",
-      "dom",
-      "dom.iterable"
-    ],
-    "target": "es6",
-    "jsx": "preserve",
-    "moduleResolution": "node",
-    "strict": true,
-    "skipLibCheck": true,
-    "allowSyntheticDefaultImports": true,
-    "noEmit": true
-  }
-}
