+++ application/extensions/bulma-richtext/.eslintignore
