+++ application/extensions/bulma-richtext/.eslintrc.js
