+++ application/extensions/bulma-richtext/.mvn/wrapper/maven-wrapper.properties
