+++ application/extensions/bulma-richtext/README.md
@@ -1,0 +1,5 @@
+# RTE extensions
+
+## Color extension
+
+## Code extension
