+++ application/extensions/bulma-richtext/babel/.babelrc.js
