+++ application/extensions/bulma-richtext/bnd.bnd
@@ -1,0 +1,1 @@
+Sling-Bundle-Resources: /libs/extensions/dialogs/components/richtext;/libs/bulma-richtext
