+++ application/extensions/bulma-richtext/src/main/webapp/libs/bulma-color/web-resources/components/richtext/plugin/extension-color.ts
