+++ application/extensions/pom.xml
@@ -33,8 +33,7 @@
     <modules>
     <module>bulma-table-service</module>
     <module>bulma-table-view</module>
-    <module>bulma-color</module>
-    <module>bulma-code</module>
+    <module>bulma-richtext</module>
     </modules>
 
     <properties>
