+++ distribution/pom.xml
@@ -63,13 +63,7 @@
     </dependency>
     <dependency>
       <groupId>pl.ds.bulma</groupId>
-      <artifactId>bulma-color</artifactId>
-      <version>1.2.3-SNAPSHOT</version>
-      <type>pom</type>
-    </dependency>
-    <dependency>
-      <groupId>pl.ds.bulma</groupId>
-      <artifactId>bulma-code</artifactId>
+      <artifactId>bulma-richtext</artifactId>
       <version>1.2.3-SNAPSHOT</version>
       <type>pom</type>
     </dependency>
